requirements.txt
absl-py==0.6.1
antiorm==1.2.1
astor==0.7.1
astroid==2.2.5
autopep8==1.4.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
db==0.1.1
db-sqlite3==0.0.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.3
django-allauth==0.39.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.4
Flask==1.0.3
gast==0.2.0
grpcio==1.16.1
gunicorn==19.9.0
h5py==2.8.0
httpie==1.0.2
idna==2.5
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.4.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.13
mysqlclient==1.3.12
newsapi-python==0.2.3
numpy==1.15.4
oauthlib==3.0.2
openapi-codec==1.3.2
pep8==1.7.1
pip-lock==2.0.0
pipenv==2018.11.26
protobuf==3.6.1
psycopg2==2.8.3
pycodestyle==2.5.0
Pygments==2.4.2
pylint==2.3.1
python3-openid==3.1.0
pytz==2018.7
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.21.1
virtualenv==16.1.0
virtualenv-clone==0.5.3
Werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.11.2